Trait effective::Effective

pub trait Effective {
    type Output;
    type Residual;
    type Yields: Exists;
    type Awaits: Exists;

    // Required method
    fn poll_effect(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>
    ) -> EffectResult<Self::Output, Self::Residual, Self::Yields, Self::Awaits>;
}

Effective encapsulates all possible effect types that Rust currently has. Fallability, Iterability and Awaitablilty.

Required Associated Types

type Output

What item does this effective type produce

type Residual

What non-success types can this effective produce

type Yields: Exists

Models whether this effective type can yield multiple values

type Awaits: Exists

Models whether this effective type can await

Required Methods

fn poll_effect( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> EffectResult<Self::Output, Self::Residual, Self::Yields, Self::Awaits>