Trait effective::Effective

pub trait Effective {
    type Item: Try;
    type Yields: Exists;
    type Awaits: Exists;

    // Required method
    fn poll_effect(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>
    ) -> EffectResult<Self::Item, Self::Yields, Self::Awaits>;
}

Effective encapsulates all possible effect types that Rust currently has. Fallability, Iterability and Awaitablilty.

Required Associated Types

type Item: Try

Models how this effective type can fail.

type Yields: Exists

Models whether this effective type can yield multiple values

type Awaits: Exists

Models whether this effective type can await

Required Methods

fn poll_effect( self: Pin<&mut Self>, cx: &mut Context<'_> ) -> EffectResult<Self::Item, Self::Yields, Self::Awaits>